Moved show_*_dialog function to new helper module
